requirements.txt
requests==2.31.0
web3==6.15.1
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
